Moved client-side model type definitions to a separate file
