+++ pages/index.tsx
@@ -3,15 +3,10 @@ import styled from 'react-emotion';
 import withDefault, { DefaultProps } from './_app';
 import { Banner, colors, Agenda } from 'p20-components';
 import { AgendaType } from 'p20-components/types/Agenda';
-import {
-  SpeakerCategory,
-  Speaker as SpeakerModel,
-  SponsorCategory as SponsorCategoryModel,
-  Sponsor as SponsorModel
-} from '../models';
 import Tickets from '../components/Tickets';
 import Speakers from '../components/Speakers';
 import Sponsors from '../components/Sponsors';
+import { SpeakerCategories, SponsorCategories } from '../components/types';
 
 interface Props {
   speakerCategories: SpeakerCategories;
@@ -20,7 +15,7 @@ interface Props {
   agenda: AgendaType;
 }
 
-const Background = styled('section')({
+export const Background = styled('section')({
   '& > *:nth-child(odd)': {
     backgroundColor: `${colors.lightGray}`
   },
@@ -28,12 +23,6 @@ const Background = styled('section')({
     backgroundColor: `${colors.white}`
   }
 });
-
-type MarkdownDescription = { description: { md: string } };
-export type Speaker = (SpeakerModel & MarkdownDescription);
-export type SpeakerCategories = (SpeakerCategory & { speakers: Speaker[] })[];
-export type Sponsor = (SponsorModel & MarkdownDescription);
-export type SponsorCategories = (SponsorCategoryModel & { sponsors: Sponsor[] })[];
 
 const Home: React.StatelessComponent<DefaultProps & Props> = ({
   agenda,
